Changed timeing code

+++ Launcher.c
@@ -6,10 +6,10 @@ int main(int argc, char * argv[]) {
     printf("----------------------------------------------------------------------------");
     printf("\nAndBench Started...");
     printf("\n----------------------------------------------------------------------------\n");
-    int cpuScore = cpu_calculate(argc,argv);
-    printf(" CPU Score : %d",cpuScore);
-    int fileWriteScore = writeTest();
-    printf("\n Write Speed Score : %d",fileWriteScore);
-    int fileReadScore =readTest();
-    printf("\n Read Speed Score : %d\n",fileReadScore);
+    unsigned long cpuScore = cpu_calculate(argc,argv);
+    printf(" CPU Score : %lu",cpuScore);
+    unsigned long fileWriteScore = writeTest();
+    printf("\n Write Speed Score : %lu",fileWriteScore);
+    unsigned long fileReadScore =readTest();
+    printf("\n Read Speed Score : %lu\n",fileReadScore);
 }
